BACK_END/pyproject.toml
[tool.poetry]
name = "projeto crud"
version = "1.0.0"
description = ""
authors = ["Raivaz <rai.vaz@inacsistemas.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-dotenv = "^1.0.1"
mysql-connector-python = "^8.3.0"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
